Another pass to clean up Sonar and compiler warnings for maintainability

See merge request weblogic-cloud/weblogic-kubernetes-operator!4949

+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItLBTwoDomainsTraefik.java
@@ -224,13 +224,11 @@ class ItLBTwoDomainsTraefik {
    */
   @AfterAll
   void tearDownAll() throws ApiException {
-    if (!SKIP_CLEANUP) {
-      if (pvPvcNamePair != null) {
-        // delete pvc
-        deletePersistentVolumeClaim(pvPvcNamePair.get(1), domainNamespace);
-        // delete pv
-        deletePersistentVolume(pvPvcNamePair.get(0));
-      }
+    if (!SKIP_CLEANUP && pvPvcNamePair != null) {
+      // delete pvc
+      deletePersistentVolumeClaim(pvPvcNamePair.get(1), domainNamespace);
+      // delete pv
+      deletePersistentVolume(pvPvcNamePair.get(0));
     }
     if (traefikHelmParams != null && OKE_CLUSTER) {
